refactor(#10): move buttons API to another dir
